requirements.txt
# requirements.txt
streamlit
Pillow
numpy
torch
torchvision
transformers
spacy
timm
# https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1.tar.gz